+++ tests/e2e/cases/generated_c_dce_const/in.ts
@@ -1420,11 +1420,15 @@ const unused_error_descriptor_value_upper_call = Object.getOwnPropertyDescriptor
 const unused_error_descriptors_value_upper_call = Object.getOwnPropertyDescriptors(new TypeError("dead_error_descriptors_value_upper_call")).message.value.toUpperCase();
 const unused_error_descriptor_reflect_get_value_upper_call = Reflect.get(Object.getOwnPropertyDescriptor(new Error("dead_error_descriptor_reflect_get_value_upper"), "message")!, "value").toUpperCase();
 const unused_error_descriptor_reflect_get_value_receiver_upper_call = Reflect.get(Object.getOwnPropertyDescriptor(new Error("dead_error_descriptor_reflect_get_value_receiver_upper"), "message")!, "value", { dead_error_descriptor_reflect_get_value_receiver: 1 }).toUpperCase();
+// @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+const unused_error_descriptor_nested_descriptor_value_upper_call = Object.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(new Error("dead_error_descriptor_nested_descriptor_value_upper"), "message")!, "value", "dead_error_descriptor_nested_descriptor_ignored")!.value.toUpperCase();
 const unused_regexp_descriptor_value_read = Object.getOwnPropertyDescriptor(/dead_regexp_descriptor_value_read/g, "lastIndex")!.value;
 const unused_regexp_descriptors_value_read = Object.getOwnPropertyDescriptors(new RegExp("dead_regexp_descriptors_value_read", "g")).lastIndex.value;
 const unused_regexp_descriptor_value_to_string_call = Object.getOwnPropertyDescriptor(/dead_regexp_descriptor_value_to_string_call/g, "lastIndex")!.value.toString().length;
 const unused_regexp_descriptors_value_to_string_call = Object.getOwnPropertyDescriptors(new RegExp("dead_regexp_descriptors_value_to_string_call", "g")).lastIndex.value.toLocaleString().length;
 const unused_regexp_descriptor_reflect_get_value_to_string_call = Reflect.get(Object.getOwnPropertyDescriptor(/dead_regexp_descriptor_reflect_get_value_to_string/g, "lastIndex")!, "value").toString().length;
+// @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+const unused_regexp_descriptor_nested_reflect_descriptor_value_to_string_call = Reflect.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(/dead_regexp_descriptor_nested_reflect_descriptor_value_to_string/g, "lastIndex")!, "value", "dead_regexp_descriptor_nested_reflect_descriptor_ignored")!.value.toString().length;
 const unused_error_descriptors_keys_with_call = Object.keys(Object.getOwnPropertyDescriptors(new Error("dead_error_descriptors_keys_with"))).with(3, "dead_error_descriptors_keys_with_replacement");
 const unused_regexp_descriptors_values_with_call = Object.values(Object.getOwnPropertyDescriptors(/dead_regexp_descriptors_values_with/g)).with(0, { value: 1 });
 const unused_date_descriptors_own_keys_empty_map_call = Reflect.ownKeys(Object.getOwnPropertyDescriptors(new Date("2020-01-13T14:15:16Z"))).map(() => "dead_date_descriptors_own_keys_empty_map");
@@ -2876,11 +2880,15 @@ Object.getOwnPropertyDescriptor(new Error("top_level_dead_error_descriptor_value
 Object.getOwnPropertyDescriptors(new TypeError("top_level_dead_error_descriptors_value_upper_call")).message.value.toUpperCase();
 Reflect.get(Object.getOwnPropertyDescriptor(new Error("top_level_dead_error_descriptor_reflect_get_value_upper"), "message")!, "value").toUpperCase();
 Reflect.get(Object.getOwnPropertyDescriptor(new Error("top_level_dead_error_descriptor_reflect_get_value_receiver_upper"), "message")!, "value", { top_level_dead_error_descriptor_reflect_get_value_receiver: 1 }).toUpperCase();
+// @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+Object.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(new Error("top_level_dead_error_descriptor_nested_descriptor_value_upper"), "message")!, "value", "top_level_dead_error_descriptor_nested_descriptor_ignored")!.value.toUpperCase();
 Object.getOwnPropertyDescriptor(/top_level_dead_regexp_descriptor_value_read/g, "lastIndex")!.value;
 Object.getOwnPropertyDescriptors(new RegExp("top_level_dead_regexp_descriptors_value_read", "g")).lastIndex.value;
 Object.getOwnPropertyDescriptor(/top_level_dead_regexp_descriptor_value_to_string_call/g, "lastIndex")!.value.toString().length;
 Object.getOwnPropertyDescriptors(new RegExp("top_level_dead_regexp_descriptors_value_to_string_call", "g")).lastIndex.value.toLocaleString().length;
 Reflect.get(Object.getOwnPropertyDescriptor(/top_level_dead_regexp_descriptor_reflect_get_value_to_string/g, "lastIndex")!, "value").toString().length;
+// @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+Reflect.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(/top_level_dead_regexp_descriptor_nested_reflect_descriptor_value_to_string/g, "lastIndex")!, "value", "top_level_dead_regexp_descriptor_nested_reflect_descriptor_ignored")!.value.toString().length;
 Object.keys(Object.getOwnPropertyDescriptors(new Error("top_level_dead_error_descriptors_keys_with"))).with(3, "top_level_dead_error_descriptors_keys_with_replacement");
 Object.values(Object.getOwnPropertyDescriptors(/top_level_dead_regexp_descriptors_values_with/g)).with(0, { value: 1 });
 Reflect.ownKeys(Object.getOwnPropertyDescriptors(new Date("2020-01-14T15:16:17Z"))).map(() => "top_level_dead_date_descriptors_own_keys_empty_map");
@@ -4375,11 +4383,15 @@ function usedLocal(value: number): number {
     Object.getOwnPropertyDescriptors(new TypeError("local_dead_error_descriptors_value_upper_call")).message.value.toUpperCase();
     Reflect.get(Object.getOwnPropertyDescriptor(new Error("local_dead_error_descriptor_reflect_get_value_upper"), "message")!, "value").toUpperCase();
     Reflect.get(Object.getOwnPropertyDescriptor(new Error("local_dead_error_descriptor_reflect_get_value_receiver_upper"), "message")!, "value", { local_dead_error_descriptor_reflect_get_value_receiver: 1 }).toUpperCase();
+    // @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+    Object.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(new Error("local_dead_error_descriptor_nested_descriptor_value_upper"), "message")!, "value", "local_dead_error_descriptor_nested_descriptor_ignored")!.value.toUpperCase();
     Object.getOwnPropertyDescriptor(/local_dead_regexp_descriptor_value_read/g, "lastIndex")!.value;
     Object.getOwnPropertyDescriptors(new RegExp("local_dead_regexp_descriptors_value_read", "g")).lastIndex.value;
     Object.getOwnPropertyDescriptor(/local_dead_regexp_descriptor_value_to_string_call/g, "lastIndex")!.value.toString().length;
     Object.getOwnPropertyDescriptors(new RegExp("local_dead_regexp_descriptors_value_to_string_call", "g")).lastIndex.value.toLocaleString().length;
     Reflect.get(Object.getOwnPropertyDescriptor(/local_dead_regexp_descriptor_reflect_get_value_to_string/g, "lastIndex")!, "value").toString().length;
+    // @ts-ignore: intentional ignored-argument coverage for generated-C DCE.
+    Reflect.getOwnPropertyDescriptor(Object.getOwnPropertyDescriptor(/local_dead_regexp_descriptor_nested_reflect_descriptor_value_to_string/g, "lastIndex")!, "value", "local_dead_regexp_descriptor_nested_reflect_descriptor_ignored")!.value.toString().length;
     Object.keys(Object.getOwnPropertyDescriptors(new Error("local_dead_error_descriptors_keys_with"))).with(3, "local_dead_error_descriptors_keys_with_replacement");
     Object.values(Object.getOwnPropertyDescriptors(/local_dead_regexp_descriptors_values_with/g)).with(0, { value: 1 });
     Reflect.ownKeys(Object.getOwnPropertyDescriptors(new Date("2020-01-15T16:17:18Z"))).map(() => "local_dead_date_descriptors_own_keys_empty_map");